// packages/client-expo/src/types.ts
export interface Config<ThrowOnError extends boolean = boolean>
	extends Omit<FetchRequestInit, "body" | "headers" | "method"> {
	/**
	 * Base URL for all requests made by this client.
	 * @default ''
	 */
	baseUrl?: string;
	/**
	 * Any body that you want to add to your request.
	 *
	 * {@link https://developer.mozilla.org/docs/Web/API/fetch#body}
	 */
	body?:
		| FetchRequestInit["body"]
		| Record<string, unknown>
		| Array<Record<string, unknown>>
		| Array<unknown>
		| number;
	/**
	 * A function for serializing request body parameter. By default,
	 * {@link JSON.stringify()} will be used.
	 */
	bodySerializer?: BodySerializer;
	/**
	 * Fetch API implementation. You can use this option to provide a custom
	 * fetch instance.
	 * @default globalThis.fetch
	 */
	fetch?: (url: string, init?: FetchRequestInit) => ReturnType<typeof fetch>;
	/**
	 * An object containing any HTTP headers that you want to pre-populate your
	 * `Headers` object with.
	 *
	 * {@link https://developer.mozilla.org/docs/Web/API/Headers/Headers#init See more}
	 */
	headers?:
		| RequestInit["headers"]
		| Record<
				string,
				| string
				| number
				| boolean
				| (string | number | boolean)[]
				| null
				| undefined
				| unknown
		  >;
	/**
	 * The request method.
	 *
	 * {@link https://developer.mozilla.org/docs/Web/API/fetch#method See more}
	 */
	method?:
		| "CONNECT"
		| "DELETE"
		| "GET"
		| "HEAD"
		| "OPTIONS"
		| "PATCH"
		| "POST"
		| "PUT"
		| "TRACE";
	/**
	 * Return the response data parsed in a specified format. By default, `auto`
	 * will infer the appropriate method from the `Content-Type` response header.
	 * You can override this behavior with any of the {@link Body} methods.
	 * Select `stream` if you don't want to parse response data at all.
	 * @default 'auto'
	 */
	parseAs?: Exclude<keyof Body, "body" | "bodyUsed"> | "auto" | "stream";
	/**
	 * A function for serializing request query parameters. By default, arrays
	 * will be exploded in form style, objects will be exploded in deepObject
	 * style, and reserved characters are percent-encoded.
	 *
	 * {@link https://swagger.io/docs/specification/serialization/#query View examples}
	 */
	querySerializer?: QuerySerializer | QuerySerializerOptions;
	/**
	 * A function for transforming response data before it's returned to the
	 * caller function. This is an ideal place to post-process server data,
	 * e.g. convert date ISO strings into native Date objects.
	 */
	responseTransformer?: (data: unknown) => Promise<unknown>;
	/**
	 * Throw an error instead of returning it in the response?
	 * @default false
	 */
	throwOnError?: ThrowOnError;
}

export interface RequestOptionsBase<ThrowOnError extends boolean>
	extends Config<ThrowOnError> {
	path?: Record<string, unknown>;
	query?: Record<string, unknown>;
	url: string;
}

export type RequestResult<
	Data = unknown,
	TError = unknown,
	ThrowOnError extends boolean = boolean,
> = ThrowOnError extends true
	? Promise<{
			data: Data;
			request: Request;
			response: Response;
		}>
	: Promise<
			(
				| { data: Data; error: undefined }
				| { data: undefined; error: TError }
			) & {
				request: Request;
				response: Response;
			}
		>;

type MethodFn = <
	Data = unknown,
	TError = unknown,
	ThrowOnError extends boolean = false,
>(
	options: Omit<RequestOptionsBase<ThrowOnError>, "method">,
) => RequestResult<Data, TError, ThrowOnError>;

type RequestFn = <
	Data = unknown,
	TError = unknown,
	ThrowOnError extends boolean = false,
>(
	options: Omit<RequestOptionsBase<ThrowOnError>, "method"> &
		Pick<Required<RequestOptionsBase<ThrowOnError>>, "method">,
) => RequestResult<Data, TError, ThrowOnError>;

export interface Client {
	connect: MethodFn;
	delete: MethodFn;
	get: MethodFn;
	getConfig: () => Config;
	head: MethodFn;
	options: MethodFn;
	patch: MethodFn;
	post: MethodFn;
	put: MethodFn;
	request: RequestFn;
	setConfig: (config: Config) => Config;
	trace: MethodFn;
}

export type RequestOptions = RequestOptionsBase<false> &
	Config<false> & {
		headers: Headers;
	};